// apps/website/components/features/landing/types/landing.types.ts
export interface IFeature {
  number: string
  title: string
  description: string
}

export interface IStep {
  number: string
  title: string
  description: string
}

export interface ISecurityBullet {
  title: string
  description: string
}

export interface IFaq {
  question: string
  answer: string
}